requirements.txt
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
environs==14.2.0
Markdown==3.8
pillow==11.1.0
tinycss2==1.4.0

